테스트 케이스 추가 및 수정

+++ src/main/java/meeting/decision/service/RoomService.java
@@ -10,10 +10,9 @@ import meeting.decision.dto.room.RoomOutDTO;
 import meeting.decision.dto.room.RoomUpdateDTO;
 import meeting.decision.dto.user.UserOutDTO;
 import meeting.decision.exception.DuplicateUserExeption;
+import meeting.decision.exception.RoomNotFoundException;
 import meeting.decision.exception.UserNotFoundErrorException;
-import meeting.decision.repository.JpaRoomParticipantRepository;
-import meeting.decision.repository.JpaRoomRepository;
-import meeting.decision.repository.JpaUserRepository;
+import meeting.decision.repository.*;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -31,6 +30,8 @@ public class RoomService {
     private final JpaRoomRepository roomRepository;
     private final JpaUserRepository userRepository;
     private final JpaRoomParticipantRepository roomParticipantRepository;
+    private final JpaVotePaperRepository votePaperRepository;
+    private final JpaVoteRepository voteRepository;
 
 
     //create
@@ -45,7 +46,7 @@ public class RoomService {
 
     @CheckUser(isOwner = true)
     public void update(Long ownerId, Long roomId, RoomUpdateDTO updateParam){ //owner인지 확인하는거 aop로 적용 Parameter에 Long 두개 있는지 확인
-        Room room = roomRepository.findById(roomId).orElseThrow();
+        Room room = roomRepository.findById(roomId).orElseThrow(RoomNotFoundException::new);
         room.setRoomName(updateParam.getRoomName());
         room.setOwner(updateParam.getOwner());
     }
@@ -53,31 +54,32 @@ public class RoomService {
     //remove
     @CheckUser(isOwner = true)
     public void delete(Long ownerId, Long roomId){
+        //cascade.REMOVE 가 삭제를 해주지만 너무 많은 DELETE 쿼리가 나가므로 직접 DELETE 문으로 삭제
+        votePaperRepository.deleteVotePaperByRoomId(roomId);
+        voteRepository.deleteVoteByRoomId(roomId);
+        roomParticipantRepository.deleteByRoomId(roomId);
         roomRepository.deleteById(roomId);
     }
 
     //delete user
     @CheckUser(isOwner = true)
     public void deleteUserFromRoom(Long ownerId, Long roomId, Long userId){
-        Room room = roomRepository.findById(roomId).orElseThrow();
         User user = userRepository.findById(userId).orElseThrow();
         if(user.getId().equals(ownerId)){   //Owner가 삭제될 경우 Room이 삭제
             delete(ownerId, roomId); // 내부 호출 주의
             return;
         }
 
-    roomParticipantRepository.deleteByRoomIdAndUserId(roomId, userId);
-//        Optional<RoomParticipant> participant = roomParticipantRepository.findByRoomIdAndUserId(roomId, user.getId());
-//        if(participant.isEmpty()){
-//            throw new UserNotFoundErrorException();
-//        }
-//
-//        //Lazy기 때문에 Select Qeury가 나감. 뒤에 더 작업을 안하니 지워줌. ?????? 이걸 해주는게 맞을지 아닐지 .... 고민을 해봅시다...
+        Optional<RoomParticipant> participant = roomParticipantRepository.findByRoomIdAndUserId(roomId, user.getId());
+        if(participant.isEmpty()){
+            return;
+        }
+
+        roomParticipantRepository.deleteByRoomIdAndUserId(roomId, userId);
+
+//        //Lazy기 때문에 Select Qeury가 나감.
 //        room.getUserList().remove(participant.get());
 //        user.getRoomList().remove(participant.get());
-//
-//        roomParticipantRepository.delete(participant.get());
-
     }
     //내일 이거 테이블 바꿔보기
     //add User에서 select join이 너무 많이 일어나는 문제가 발생
@@ -85,7 +87,7 @@ public class RoomService {
     @CheckUser(isOwner = true)
     public void addUserToRoom(Long ownerId, Long roomId, Long userId){
 
-        Room room = roomRepository.findById(roomId).orElseThrow();
+        Room room = roomRepository.findById(roomId).orElseThrow(RoomNotFoundException::new);
         User user = userRepository.findById(userId).orElseThrow();
 
         if(roomParticipantRepository.existsByRoomIdAndUserId(roomId, userId)){
@@ -96,7 +98,7 @@ public class RoomService {
     }
 
     public List<RoomParticipant> findAllUser(Long roomId){
-        return roomRepository.findById(roomId).orElseThrow().getUserList();
+        return roomRepository.findById(roomId).orElseThrow(RoomNotFoundException::new).getUserList();
     }
 
     public List<Room> findAll(){
@@ -105,13 +107,12 @@ public class RoomService {
 
     @CheckUser
     public List<UserOutDTO> findAllUserByRoomId(Long userId, Long roomId){
-        List<UserOutDTO> list = new ArrayList<>();
-        return roomRepository.findById(roomId).orElseThrow().getUserList().stream().map((participant -> new UserOutDTO(participant.getUser().getId(), participant.getUser().getUsername()))).collect(Collectors.toList());
+        return roomRepository.findById(roomId).orElseThrow(RoomNotFoundException::new).getUserList().stream().map((participant -> new UserOutDTO(participant.getUser().getId(), participant.getUser().getUsername()))).collect(Collectors.toList());
     }
 
     @CheckUser
     public RoomOutDTO findRoomById(Long userId, Long roomId){
-        return roomParticipantRepository.findRoomOutDTOByRoomId(roomId).orElseThrow();
+        return roomParticipantRepository.findRoomOutDTOByRoomId(roomId).orElseThrow(RoomNotFoundException::new);
     }
